Continue simple Calculator
changes:
- change int to double in parameter
- add enum for operator
- rename Menu.cs to Calculator.cs
- add user input
- add operator implementation
- add some todo to execute next iteration

+++ Learn_CSharp_Master_Class/Simple_Calculator/Interface/IOperator.cs
@@ -2,11 +2,11 @@
 
 public interface IOperator
 {
-    double Add(int firstNum, int secondNum);
+    double Add(double firstNum, double secondNum);
 
-    double Substract(int firstNum, int secondNum);
+    double Substract(double firstNum, double secondNum);
 
-    double Multiply(int firstNum, int secondNum);
+    double Multiply(double firstNum, double secondNum);
 
-    double Divide(int firstNum, int secondNum);
+    double Divide(double firstNum, double secondNum);
 }
